Fix String::irtrim and warning in File

+++ source/File.cpp
@@ -21,7 +21,7 @@ File::ReadResult File::read_line(bool& eof_reached) {
             // found eol, seek file pointer to just after it
             int64_t offset = -(static_cast<int64_t>(read) - static_cast<int64_t>(aftereolidx));
             auto old_pos   = ARLib::ftell(m_ptr);
-            auto new_pos   = ARLib::fseek(m_ptr, offset, SEEK_CUR);
+            auto new_pos   = ARLib::fseek(m_ptr, static_cast<long>(offset), SEEK_CUR);
             HARD_ASSERT(old_pos - new_pos == -offset, "pebnis");
             line += move(partial);
             break;
